Лабораторная работа №2, Шпанцев Андрей, ПИБД-21
